Made changes to work on linux, fixed several bugs in image array class

+++ C++/fractal/include/imageArray.h
@@ -8,9 +8,9 @@ class ImageArray //A class that handles a 2D array of ints.
 {
     private:
         const int MAX_VAL = 65536;
-        int *histogram;
+        unsigned int *histogram;
         void updateHistogram();
-        int findMax();
+        unsigned short findMax();
 
     public:
         const int WIDTH;
@@ -18,7 +18,7 @@ class ImageArray //A class that handles a 2D array of ints.
 
         struct complexPixel //incase we want to add extra info for each point
         {
-            short iters;
+            unsigned short iters;
         };
 
         complexPixel **pixelMatrix;
